// ui/src/api/platform-management/client/api.ts
export interface Credential {
    /**
     * 
     * @type {string}
     * @memberof Credential
     */
    'id': string;
    /**
     * 
     * @type {Array<string>}
     * @memberof Credential
     */
    'scopesRaw': Array<string>;
    /**
     * 
     * @type {CredentialData}
     * @memberof Credential
     */
    'data': CredentialData;
}
/**
 * 
 * @export
 * @interface CredentialCreateRequest
 */
export interface CredentialCreateRequest {
    /**
     * 
     * @type {Array<CredentialEntityLinkCreateDTO>}
     * @memberof CredentialCreateRequest
     */
    'entityLinks': Array<CredentialEntityLinkCreateDTO>;
    /**
     * 
     * @type {Credential}
     * @memberof CredentialCreateRequest
     */
    'credential': Credential;
    /**
     * 
     * @type {string}
     * @memberof CredentialCreateRequest
     */
    'fullPathOwnerGroupId': string;
}
/**
 * @type CredentialData
 * @export
 */
export type CredentialData = CredentialDataKeyPair | CredentialDataUsernamePassword;

/**
 * 
 * @export
 * @interface CredentialDataKeyPair
 */
export interface CredentialDataKeyPair extends CredentialData {
    /**
     * 
     * @type {string}
     * @memberof CredentialDataKeyPair
     */
    'privateKey': string;
    /**
     * 
     * @type {string}
     * @memberof CredentialDataKeyPair
     */
    'publicKey': string;
}
// ...
export interface CredentialDataUsernamePassword extends CredentialData {
    /**
     * 
     * @type {string}
     * @memberof CredentialDataUsernamePassword
     */
    'username': string;
    /**
     * 
     * @type {string}
     * @memberof CredentialDataUsernamePassword
     */
    'password': string;
}


/**
 * 
 * @export
 * @interface CredentialDataUsernamePasswordReadDTO
 */
export interface CredentialDataUsernamePasswordReadDTO extends CredentialDataReadDTO {
    /**
     * 
     * @type {string}
     * @memberof CredentialDataUsernamePasswordReadDTO
     */
    'username': string;
}


/**
 * 
 * @export
 * @interface CredentialEntityLinkCreateDTO
 */
export interface CredentialEntityLinkCreateDTO {
    /**
     * 
     * @type {string}
     * @memberof CredentialEntityLinkCreateDTO
     */
    'entityType': string;
    /**
     * 
     * @type {string}
     * @memberof CredentialEntityLinkCreateDTO
     */
    'entityId': string;
}